cpu/gdt: remove needless cast
Fixes the following clippy warning:

warning: casting to the same type is unnecessary (`u64` -> `u64`)
  --> src/cpu/gdt.rs:45:14
   |
45 |     desc1 |= (addr >> 32) as u64;
   |              ^^^^^^^^^^^^^^^^^^^ help: try: `(addr >> 32)`
   |
   = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast
   = note: `#[warn(clippy::unnecessary_cast)]` on by default

+++ src/cpu/gdt.rs
@@ -42,7 +42,7 @@ pub fn load_tss(tss: &X86Tss) {
     // Address
     desc0 |= (addr & 0x00ff_ffffu64) << 16;
     desc0 |= (addr & 0xff00_0000u64) << 32;
-    desc1 |= (addr >> 32) as u64;
+    desc1 |= addr >> 32;
 
     // Present
     desc0 |= 1u64 << 47;
